move the table package out of cmd package because is more broadly used
